// frontend/src/types/doctor.ts
import { OfficeHours } from "./officeHours"
import { Specialization } from "./specialization"
import { Specialty } from "./specialty"

export type DoctorResponse = {
    content: Doctor[];
    last: boolean;
    totalElements: number;
    totalPages: number;
    size?: number;
    number: number;
    first: boolean;
    numberOfElements?: number;
    empty?: boolean;
}

export type Doctor = {
    id: number;
    imgUrl: string;
    crm: string;
    name: string;
    cardName: string;
    phone: string;
    email: string;
    birthDate: string;
    resume: string;
    specialty: Specialty;
    officeHours: OfficeHours;
    specializations: Specialization[];
}